requirements.txt
paho-mqtt==2.1.0
pynput==1.7.7
six==1.16.0
python-dotenv~=1.0.1